moving includes around to remove vec2 from the standard includes
