+++ Application/src/commons/common/misc/Grid.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <types.h>
+#include <misc/vec2.h>
 
 namespace cmn {
 namespace grid {
