+++ Application/src/commons/common/processing/PadImage.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <types.h>
+#include <misc/vec2.h>
 
 namespace cmn {
     /*template<int twidth, int theight, uint depth = 1>
